+++ src/app/(bg)/layout.tsx
@@ -1,0 +1,6 @@
+'use client'
+import { BgBody } from '@/components'
+
+export default function BgLayout({ children }: { children: React.ReactNode }) {
+  return <BgBody>{children}</BgBody>
+}
