+++ src/app/(bg)/rsa/calc.ts
