+++ src/components/background/index.ts
@@ -1,0 +1,1 @@
+export { BgBody } from './background'
