+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export { MouseStalker } from './mouse-stalker'
+export { BgBody } from './background'
